+++ app/layout.tsx
@@ -1,8 +1,8 @@
 import type { Metadata } from "next";
 
 export const metadata: Metadata = {
-  title: "SD Core",
-  description: "SD Core NMS",
+  title: "Aether SD-Core",
+  description: "Aether SD-Core NMS",
 };
 
 export default function RootLayout({
